requirements.txt
mtcnn==0.1.0
# tensorflow==2.3.1

keras-vggface==0.6
keras_applications==1.0.8
PyYAML
tqdm
scikit-learn
streamlit
bing-image-downloader

#local packages-
-e .

# Core requirements
opencv-python-headless==4.5.5.64
tensorflow==2.10.0
keras==2.10.0


streamlit==1.12.0
numpy==1.23.5
Pillow==9.3.0

# Pinned versions from your logs
Pygments==2.19.2
mdurl==0.1.2
markdown-it-py==3.0.0
rich==14.1.0